feat(TopBottomPresentable): Adds 'TopBottomPresentable' protocol and conformance on existing 'Animators' to allow '.top/.bottom' presentation w/ PhysicsAnimator
