+++ SwiftMessages/UIViewController+Extensions.swift
@@ -85,17 +85,3 @@ extension UIViewController {
         return true
     }
 }
-
-extension SwiftMessages.PresentationStyle {
-    /// A temporary workaround to allow custom presentation contexts using `TopBottomAnimation`
-    /// to display properly behind bars. THe long term solution is to refactor all of the
-    /// presentation context logic to work with safe area insets.
-    var topBottomStyle: TopBottomAnimation.Style? {
-        switch self {
-        case .top: return .top
-        case .bottom: return .bottom
-        case .custom(let animator): return (animator as? TopBottomAnimation)?.style
-        case .center: return nil
-        }
-    }
-}
